Tidied up search routes
